fix(concrete_fftw_sys): fixes broken crates (#8)
The CI of the concrete repository was broken because of issues in
the concrete-fftw crates, which are fixed by this commit.

Using `thiserror` as a dependency brings the `syn` crate in the
dependencies, which is incompatible with the `-Cpanic=abort` rustflag.
This rustflag was necessary to compute code coverage though. This
dependency removed this dependency and standard error management was
implemented.

Also, the folder containing the sources of fftw contianed a gitignore
file which for some reasons, made the compilation fail with the avx
flags. The gitignore file was removed, making the compilation operational
again.

Moreover, it was noticed that when `concrete-fftw` was not depending on the
local version of `concrete-fftw-sys`, the CI was not even building the
current version of `concrete-fftw-sys`. This was the main reason why
the bugs were only caught in the concrete CI and not the concrete-fftw
CI. The CI was fixed to build all the packages in the repository.

Lastly, we synchronized the versions of `concrete-fftw` with the current
`concrete-fftw-sys` to prepare for a new release.