fix(build): specify clang's `arch` flag on macOS
This fixes 2 things:

- Cross compiling to x86_64, for example from an apple silicon enabled mac
  to x86_64 when passing the `--target=x86_64-apple-darwin` argument to cargo
- Compiling for `arm64`

What happened before on a M1 enabled mac is that compiling with
```console
cargo test --target x86_64-apple-darwin
```
Would successfully build but fails to run tests with errors like:
`archive member 'malloc.o' with length 856 is not mach-o or llvm bitcode`
because the configure script would compile the fftw's C code
for the arm target while rust targets x86.

The fix is to detect which target rust is targeting
and set clang's `arch` flags (via the `CFLAGS` makefile variable)
to the matching target value.